Add check to prevent delete of default security group

+++ nova/api/ec2/vpc.py
@@ -356,7 +356,8 @@ class VpcController(object):
             # delete default security group
             groups = quantum.list_security_groups()
             for group in groups['security_groups']:
-                if group['tenant_id'] == tenant_id:
+                if (group['tenant_id'] == tenant_id  and
+                        group['name'] != 'default'):
                     quantum.delete_security_group(group['id'])
 
             # delete default network policy
